+++ adm/app/adms/Views/confEmails/viewConfEmails.php
@@ -10,7 +10,7 @@ if(!defined('C8L6K7E')){
  if(!empty($this->data['viewConfEmails'])){
     echo "<a href='" . URLADM . "edit-conf-emails/index/" . $this->data['viewConfEmails'][0]['id'] . "'>Editar</a><br>";
     echo "<a href='" . URLADM . "edit-conf-emails-password/index/" . $this->data['viewConfEmails'][0]['id'] . "'>Editar Senha</a><br>";
-    echo "<a href='" . URLADM . "delete-conf-emails/index/" . $this->data['viewConfEmails'][0]['id'] . "'>Apagar</a><br><br>";
+    echo "<a href='" . URLADM . "delete-conf-emails/index/" . $this->data['viewConfEmails'][0]['id'] . "'onclick= 'return confirm(\"Tem certeza que deseja excluir este registro?\")'>Apagar</a><br><br>";
 
  }
  if (isset($_SESSION['msg'])) {
